[TOSS-28] Really fix the fstab problem
The fixes in 3.21 and 3.22 did not work.  Boy this is really lame.
It turns out the code in rc.nfsroot-init to create /etc/fstab if it
doesn't exist, doesn't work because fstab needs to exist in order to
get /tmp mounted, and /tmp is need to process the here-is documetn
in the script to create fstab.  Removed that.

The right approach really is to supply a default /etc/fstab, so
the initial-fstab is restored as part of the package and the post
RPM scriptlet installs it as the fstab in the image.